Complete Resume Styling
This commit finalizes the resume styling to separate the elements into the inputs and the resume preview. A 'Complete Resume' and 'Edit Resume' Button was added to toggle between the resume being completed (therefore removing the additional whitespace) and editing the resume, giving the user a bit more space to work on their additions.